Use 64 bit floating point values.

+++ src/maths.rs
@@ -1,6 +1,6 @@
 //! Standard mathematical tools.
 
-pub use std::f32::consts::PI as PI;
+pub use std::f64::consts::PI as PI;
 
 pub use colors::*;
 pub use matrices::*;
@@ -14,14 +14,14 @@ mod rays;
 mod transforms;
 mod vectors;
 
-pub const EPSILON: f32 = 0.00001;
+pub const EPSILON: f64 = 0.00001;
 
 /// Determines that two values are approximately equal.
 pub trait ApproxEq<Rhs = Self> {
   fn is_approx(&self, rhs: Rhs) -> bool;
 }
 
-impl ApproxEq for f32 {
+impl ApproxEq for f64 {
   fn is_approx(&self, rhs: Self) -> bool {
     (self - rhs).abs() < EPSILON
   }
